Previous path algorithm were moved to the GraphPath class for easier integration to the VariableDependency.
